refactor(decks): load edit page from async data sources
